returns a validated input for ship placement in ValidatedUi
