+++ lib/execute.rb
@@ -10,7 +10,7 @@ grid = Grid.new(10)
 grid_display = GridDisplay.new(grid.size)
 ui = Ui.new($stdin, $stdout, grid_display)
 ships_list = ShipsList.new
-validated_ui = ValidatedUi.new(ui, ships_list)
+validated_ui = ValidatedUi.new(ui, ships_list, grid)
 computer = Computer.new(grid, ships_list)
 new_game = Game.new(grid_display, ui, grid, ships_list, computer, validated_ui)
