+++ spec/game_spec.rb
@@ -45,7 +45,7 @@ RSpec.describe Game do
     input = StringIO.new(PLAYER1+WRONG_INPUT+FIRST_SHIP+WRONG_POSITION+POSITION1+SECOND_SHIP+POSITION2+THIRD_SHIP+POSITION3+FOURTH_SHIP+POSITION4+FIFTH_SHIP+POSITION5+SIXTH_SHIP+POSITION6+PLAYER2+
                          WRONG_ATTACK+SINK_SHIP1+SINK_SHIP2+HIT_SHIP3+SINK_SHIP3+HIT_SHIP4+SINK_SHIP4+HIT_SHIP5+HIT2_SHIP5+SINK_SHIP5+HIT_SHIP6+HIT2_SHIP6+HIT3_SHIP6+SINK_SHIP6)
     ui = Ui.new(input, output, grid_display)
-    validated_ui = ValidatedUi.new(ui, ships_list)
+    validated_ui = ValidatedUi.new(ui, ships_list, grid)
     game = Game.new(grid_display, ui, grid, ships_list, computer, validated_ui)
     game.start
